--- a/datafusion/physical-expr/src/lib.rs
+++ b/datafusion/physical-expr/src/lib.rs
@@ -24,6 +24,7 @@ pub mod datetime_expressions;
 pub mod execution_props;
 pub mod expressions;
 pub mod functions;
+pub mod hash_utils;
 pub mod math_expressions;
 mod physical_expr;
 pub mod planner;